Add content/copy support to eZPlatformSearch
This change modifies the addObject function to test if the content object being
added to the solr index is available in the database before continuing.
The legacy module content/copy triggers the addObject function before
the new content is actually in the database, causing an uncaught
exception in the search plugin. The try/catch approach seems best
because the function is not sent any information about how the
contentObject was created. The copied content is then added at a later
time by a frequent cronjob.