Switch from ezpublish.api.persistence_handler to ezpublish.api.storage_engine
Switch from `ezpublish.api.persistence_handler` to `ezpublish.api.storage_engine` in order to avoid unwanted side effects when indexing.
An example is when having a `ezbinaryfile` field. When loading the field value, `\eZ\Publish\Core\MVC\Symfony\FieldType\BinaryBase\ContentDownloadUrlGenerator` will be called in order to generate value for `BinaryFile\Value::$uri`. Problem is that using ezplatformsearch, the URI will be generated using admin siteaccess and cached as is in persistence cache.

This patch does not completely fix the problem, which is at the fieldtype level but using `ezpublish.api.storage_engine` workarounds the issue and avoids it to repeat with similar field types.